+++ .phan/plugins/DuplicateExpressionPlugin.php
@@ -81,6 +81,27 @@ class RedundantNodeVisitor extends PluginAwarePostAnalysisVisitor
     ];
 
     /**
+     * A subset of REDUNDANT_BINARY_OP_SET.
+     *
+     * These binary operations will make this plugin warn if both sides are literals.
+     */
+    const BINARY_OP_BOTH_LITERAL_WARN_SET = [
+        flags\BINARY_BOOL_AND            => true,
+        flags\BINARY_BOOL_OR             => true,
+        flags\BINARY_BOOL_XOR            => true,
+        flags\BINARY_IS_IDENTICAL        => true,
+        flags\BINARY_IS_NOT_IDENTICAL    => true,
+        flags\BINARY_IS_EQUAL            => true,
+        flags\BINARY_IS_NOT_EQUAL        => true,
+        flags\BINARY_IS_SMALLER          => true,
+        flags\BINARY_IS_SMALLER_OR_EQUAL => true,
+        flags\BINARY_IS_GREATER          => true,
+        flags\BINARY_IS_GREATER_OR_EQUAL => true,
+        flags\BINARY_SPACESHIP           => true,
+        flags\BINARY_COALESCE            => true,
+    ];
+
+    /**
      * @param Node $node
      * A binary operation node to analyze
      *
@@ -90,11 +111,14 @@ class RedundantNodeVisitor extends PluginAwarePostAnalysisVisitor
      */
     public function visitBinaryOp(Node $node)
     {
-        if (!\array_key_exists($node->flags, self::REDUNDANT_BINARY_OP_SET)) {
+        $flags = $node->flags;
+        if (!\array_key_exists($flags, self::REDUNDANT_BINARY_OP_SET)) {
             // Nothing to warn about
             return;
         }
-        if (ASTHasher::hash($node->children['left']) === ASTHasher::hash($node->children['right'])) {
+        $left = $node->children['left'];
+        $right = $node->children['right'];
+        if (ASTHasher::hash($left) === ASTHasher::hash($right)) {
             $this->emitPluginIssue(
                 $this->code_base,
                 $this->context,
@@ -102,9 +126,99 @@ class RedundantNodeVisitor extends PluginAwarePostAnalysisVisitor
                 'Both sides of the binary operator {OPERATOR} are the same: {CODE}',
                 [
                     PostOrderAnalysisVisitor::NAME_FOR_BINARY_OP[$node->flags],
-                    ASTReverter::toShortString($node->children['left']),
+                    ASTReverter::toShortString($left),
                 ]
             );
+            return;
+        }
+        if (!\array_key_exists($flags, self::BINARY_OP_BOTH_LITERAL_WARN_SET)) {
+            return;
+        }
+        if ($left instanceof Node) {
+            $left = self::resolveLiteralValue($left);
+            if ($left instanceof Node) {
+                return;
+            }
+        }
+        if ($right instanceof Node) {
+            $right = self::resolveLiteralValue($right);
+            if ($right instanceof Node) {
+                return;
+            }
+        }
+        $this->emitPluginIssue(
+            $this->code_base,
+            $this->context,
+            'PhanPluginBothLiteralsBinaryOp',
+            'Suspicious usage of a binary operator where both operands are literals. Expression: {CODE} {OPERATOR} {CODE} (result is {CODE})',
+            [
+                ASTReverter::toShortString($left),
+                PostOrderAnalysisVisitor::NAME_FOR_BINARY_OP[$flags],
+                ASTReverter::toShortString($right),
+                ASTReverter::toShortString(self::computeResultForBothLiteralsWarning($left, $right, $flags)),
+            ]
+        );
+    }
+
+    /**
+     * Compute result of a binary operator for a PhanPluginBothLiteralsBinaryOp warning
+     * @param int|string|float|bool|null $left left hand side of operation
+     * @param int|string|float|bool|null $right right hand side of operation
+     * @param int $flags
+     * @return int|string|float|bool|null
+     */
+    private static function computeResultForBothLiteralsWarning($left, $right, int $flags)
+    {
+        switch ($flags) {
+            case flags\BINARY_BOOL_AND:
+                return $left && $right;
+            case flags\BINARY_BOOL_OR:
+                return $left || $right;
+            case flags\BINARY_BOOL_XOR:
+                return $left xor $right;
+            case flags\BINARY_IS_IDENTICAL:
+                return $left === $right;
+            case flags\BINARY_IS_NOT_IDENTICAL:
+                return $left !== $right;
+            case flags\BINARY_IS_EQUAL:
+                return $left == $right;
+            case flags\BINARY_IS_NOT_EQUAL:
+                return $left != $right;
+            case flags\BINARY_IS_SMALLER:
+                return $left < $right;
+            case flags\BINARY_IS_SMALLER_OR_EQUAL:
+                return $left <= $right;
+            case flags\BINARY_IS_GREATER:
+                return $left > $right;
+            case flags\BINARY_IS_GREATER_OR_EQUAL:
+                return $left >= $right;
+            case flags\BINARY_SPACESHIP:
+                return $left <=> $right;
+            case flags\BINARY_COALESCE:
+                return $left ?? $right;
+            default:
+                return '(unknown)';
+        }
+    }
+
+    /**
+     * @return int|string|float|bool|null|Node the resolved value of $node, or $node if it could not be resolved
+     */
+    private static function resolveLiteralValue(Node $node)
+    {
+        if ($node->kind !== ast\AST_CONST) {
+            return $node;
+        }
+        // @phan-suppress-next-line PhanPartialTypeMismatchArgumentInternal
+        switch (\strtolower($node->children['name']->children['name'] ?? null)) {
+            case 'false':
+                return false;
+            case 'true':
+                return true;
+            case 'null':
+                return null;
+            default:
+                return $node;
         }
     }